refactor: change the base of the project
Removed old version of the project. Added new with more maven
dependencies and new naming convention
