+++ app/src/main/resources/application.properties
@@ -1,0 +1,9 @@
+spring.application.name=app
+spring.datasource.url=jdbc:postgresql://localhost:5232/app
+spring.datasource.username=user
+spring.datasource.password=password
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.hibernate.ddl-auto=create
+spring.main.banner-mode=off
+spring.security.user.password=user
+
